Redone political parties and upperhouse seats
Parties from between 1945-2000 are now also represented, but due to the lack of historical parties for the Progressive, Populist, and Nationalist ideologies there were improvisations.
Specifically for the Progressive ideology a party was 'revived', the party died in 1945 and had a revival attempt in 1953 but never took any seats and was disbanded. It is partly seen as the inspiration for the current progressive part D66 and it matches the ideology so that's why it was choosen.
Populist, being based on anti-establishment, didn't take place until 1951 with the Famers' Party. They starting date was pushed back to include them from the start.
Nationalist parties largely weren't present due to nazi and facism parties being banned, the closest match I could find was the Catholic National Party (A conservative party with anti-conlonial independece policies). The startdate was pushed back as well. The other natonal parties are mostly just short lived because they got banned pretty quickly irl, also explaining why there are so many of them.

Upperhouse seats were corrected to match the composition of the Senate in their respective years.