Flavour flag stuff
Added a special flag for a democratic Benelux:
The lion in the middle of the old flag partly represents the Monarch so it was kind of unfitting.

Modified Communist flag for Netherlands:
The old flag used bright orange which is related to the Dutch Orangisme political branch, which is very unfitting for a Communist Government to use.