Minor Benelux and Netherlands stuff (#269)
* Redone political parties and upperhouse seats

Parties from between 1945-2000 are now also represented, but due to the lack of historical parties for the Progressive, Populist, and Nationalist ideologies there were improvisations.
Specifically for the Progressive ideology a party was 'revived', the party died in 1945 and had a revival attempt in 1953 but never took any seats and was disbanded. It is partly seen as the inspiration for the current progressive part D66 and it matches the ideology so that's why it was choosen.
Populist, being based on anti-establishment, didn't take place until 1951 with the Famers' Party. They starting date was pushed back to include them from the start.
Nationalist parties largely weren't present due to nazi and facism parties being banned, the closest match I could find was the Catholic National Party (A conservative party with anti-conlonial independece policies). The startdate was pushed back as well. The other natonal parties are mostly just short lived because they got banned pretty quickly irl, also explaining why there are so many of them.

Upperhouse seats were corrected to match the composition of the Senate in their respective years.

* Reverted upperhouse due to possible bug, minor fix in party name

* goddamn it was probably the empyt enter/space/nextline in the localisation file, bug is gone now

* Minor name fix

* Religious division fix of the populace
Minor party renaming

* Minor party renaming, starting religion changed.

* Religious population changes for second startdate

* First step towards a communist unified Benelux.

* Communist Benelux formation event added, minor fix in parties Netherlands

* Flavour flag stuff

Added a special flag for a democratic Benelux:
The lion in the middle of the old flag partly represents the Monarch so it was kind of unfitting.

Modified Communist flag for Netherlands:
The old flag used bright orange which is related to the Dutch Orangisme political branch, which is very unfitting for a Communist Government to use.

* Added a special flag for theocratic government form for the Netherlands

Chance of them becoming one ingame is near zero, but in the case it happens their there flag is now visibly different from the other dictatorship ones.

* Redid flag because previous revision was too nationalistic.

* Redid republic falg of Benelux
Previous one was too 'busy' and not well visible ingame, new one is more straightup/streamlined and clearly indicates what it represents (a union between three nation states, without any Monarch influence).

* Added theocratic flag for the Benelux, based on republic one.

* Setting of a "desecularized" flag as mentioned in #250.