Trying to balance the economy
Added coffee, tea and sugar as everyday needs. **Nobody was consuming them. 
Moved canned food to EN, and fresh food to LN. **In most of the world fresh food is cheaper than processed food.
Moved clothes and footwear to LN. **Even the poorest people have clothes and shoes in the XXI century. Thanks China.
Moved telecomm and transport to EN. **They are meant to represent phone lines and mass transportation, so I dont know why they would be luxuries. They are needed to go to work in the first place.
Added paper as EN. **Nobody was consuming it and we all have at least toilet paper in our homes right? Also napkins and textbooks for the children.
Other changes I might forgot to mention.

This is the same for all pops. I decreased amounts for middle and lower class accordingly. 
I have tested this changes in 2 plays and the economy was much better.