Update Attack Helicopters.txt
Increased stats of attack helicopter, but made it much more expensive. It would be defense equivalent of tanks. It has 0.75 discipline and reduced org to simulate high firepower but low staying power, in contrast to infantry which is also defense oriented with more org. This means a pure helicopter army will deal high damage but will not stay long in front line.