Update Interceptor Fighter.txt
Similar to attack helicopter, Intercepter is another defense oriented unit. It defense equivalent of bomber. It is somewhat of a hybrid of back line and front line unit. It has excellent maneuverbility and recon, can shoot from backline, deal good damage in front, but is fragile. The significant buff also led to great increase in cost and supply consumption however.