Update Special Forces.txt
Special force becomes an attack oriented unit unlike other infantry. It is very expensive to build, but relatively cheap to maintain. Casualty is expensive. It is not so much meant to be a fodder like other infantry, but a specialist for infiltrating heavily fortified position with high siege and high recon.