Please read
These changes are needed for the sphere bug fix. However on its own it will NOT affect this mod, so please take it in. It will make my life easier when updating my game to preserve the change, as well as anyone else who want to try it.

So please take this in.