Update army_tech.txt
Adjusted power of doctrine tech a bit. Due to multiplicative nature of tactics and morale, the increase is exponential, meaning earlier upgrade should be less for a more linear result. The network centric one is an outlier because of how much it changes warfare. Leave it be as a flavor thing. 

In terms of gameplay it should make developed country stomp poorer country less early on.