Fixes
Added all suggested fixes. Problems keep occuring regarding the events as except for the abdication the ones in the italy txt file are not recognized as valid events, even the events that were already present are not regarded as valid. Tried moving them first all then in different pieces to a different file, changing ID and checked graphs and they are still not considered valid