Italy 50's
Added 1953 and 1958 elections and events related to them, San Marino Civil War and eventual showoff with America. The Italian Economic miracle decision was turned into an event reliant on the reconstruction decisions and event 8007751 was reworked to be more accurate. Issues persist with a few events, 8020313, 8020318 and 8020319 crash the game. 8007751, 8020314, 8020320, 8020321, 8020322 don't exist.