70-80's Rework
Finished content for the 70's, lacks only a few pieces of loc, only issues are found in the new string of code from the two coup events. Added the Seveso Disaster, elections for all the timeframe, two new terror attacks, the 77 movement, shift in the PCI from ML to Eurocommunism, Trentino autonomy and reviewed the Kidnapping of Aldo Moro with dynamic outcomes now.