Update clean_up.txt
I added stricter and clearer requirements for which type of dictatorship is ushered in when the vote franchise is set to none_voting based on ruling party ideology, citizenship and minority reforms, and war policy. This will lead to the player having more control over whether he gets a sectarian regime, nationalist dictatorship, or presidential dictatorship.