International banks change (#994)
Fixed a bug where the various international economic orgs would constantly change their policies - doing so once a day. This likely impacted performance somewhat. I also don't know if this is all such decisions that need to be modified.

Also, not sure if its a change I made and forgot about or something that was reverted, but there's a check in the file to make sure the Cold War is over before the WTO is formed.